+++ evaluation/VM/gdpr.sh
@@ -31,12 +31,17 @@ function prepare_configs() {
 # Call the parse_args function with your command-line arguments
 parse_args_and_checks "$@"
 
+if [[ "$server_connection" != "UNIX" ]]; then
+  echo "Error: server_connection must be 'UNIX' for the GDPR, VM experiment. TCP is currently not supported."
+  exit 1
+fi
+
 # compile the controller in the CVM with the appropriate encryption option
 virt-customize --add ${images_dir}/gdpr.img --smp $(nproc) --memsize 16384 \
   --run-command "cd /root/GDPRuler/controller && rm -rf build && cmake -S . -B build -D CMAKE_BUILD_TYPE=Release -D ENCRYPTION_ENABLED=$encryption && cmake --build build -j$(nproc)"
 
 # GDPR controller
-results_csv_file=${script_dir}/results/gdpr_CVM-query_mgmt_${workload_type}-encryption_$encryption-logging_$logging.csv
+results_csv_file=${script_dir}/results/gdpr_CVM-query_mgmt_${workload_type}-encryption_$encryption-logging_$logging-connection_${server_connection}.csv
 for n_clients in $clients; do
   # prepare the client configs
   prepare_configs $n_clients
@@ -53,7 +58,7 @@ for n_clients in $clients; do
         db_port=$ctl_redis_port
         db_address=$ctl_redis_address
       fi
-      echo "Starting a gdpr CVM scenario run with $n_clients clients, $db store, gdpr controller, $workload and logging set to $logging"
+      echo "Starting a gdpr CVM scenario run with $n_clients clients, $db store, gdpr controller, $workload, logging set to $logging, and server connection set to $server_connection"
       run_CVM_gdpr_experiment $n_clients $workload $db $db_address $db_port \
       $controller_address $controller_port $client_cfg $results_csv_file
       echo ""
