+++ evaluation/VM/passthrough.sh
@@ -12,12 +12,17 @@ source $script_dir/../args_and_checks.sh
 # Call the parse_args function with your command-line arguments
 parse_args_and_checks "$@"
 
+if [[ "$server_connection" != "UNIX" ]]; then
+  echo "Error: server_connection must be 'UNIX' for the passthrough, VM experiment. TCP is currently not supported."
+  exit 1
+fi
+
 # compile the controller in the CVM with the appropriate encryption option
 virt-customize --add ${images_dir}/gdpr.img --smp $(nproc) --memsize 16384 \
   --run-command "cd /root/GDPRuler/controller && rm -rf build && cmake -S . -B build -D CMAKE_BUILD_TYPE=Release -D ENCRYPTION_ENABLED=$encryption && cmake --build build -j$(nproc)"
 
 # GDPR controller
-results_csv_file=${script_dir}/results/passthrough_CVM-query_mgmt_${workload_type}-encryption_$encryption-logging_$logging.csv
+results_csv_file=${script_dir}/results/passthrough_CVM-query_mgmt_${workload_type}-encryption_$encryption-logging_$logging-connection_${server_connection}.csv
 for n_clients in $clients; do
   # copy the configs in the controller image
   virt-customize --add ${images_dir}/gdpr.img --copy-in ${script_dir}/../configs:/root
@@ -32,7 +37,7 @@ for n_clients in $clients; do
         db_port=$ctl_redis_port
         db_address=$ctl_redis_address
       fi
-      echo "Starting a passthrough CVM scenario run with $n_clients clients, $db store, passthrough controller, $workload and logging set to $logging"
+      echo "Starting a passthrough CVM scenario run with $n_clients clients, $db store, passthrough controller, $workload, logging set to $logging, and server connection set to $server_connection"
       run_CVM_passthrough_experiment $n_clients $workload $db $db_address $db_port \
       $controller_address $controller_port $client_cfg $results_csv_file
       echo ""
